Fixed a problem where pdf wasn't rerendering
Before if you loaded a pdf and then another it would work strangely
Seemingly the render wasn't getting called even though it totally
should.

For now I just force it by setting state to null and then asking for
the pdf, this is basically fine but there must be a better way to do it.